requirements.txt
embit>=0.7.0
qrcode[pil]>=7.0
Pillow>=9.0
